Re-enable automatic inference of inverse_of for autosave associations
Fixes issues: #9336, #22391, #22358, #23163, #23171, and probably more.

We weren't automatically inferring the inverse_of for autosaved associations. This appears to have been a hack in order to avoid double-save problems with nested attributes but at the expense of not being able to access the parent model during the save callback chain.

Among other things this was preventing validation of parent(s) when creating childen via nested attributes.

The correct fix is to automatically infer inverse_of for autosaved associations and to address the root issue of double queries, which is to ensure that child records saved in callbacks do not cause a duplicate save on the parent.

We do this by wrapping autosaves of associated records in a block which temporarily disables the reverse autosave leading back to the record which triggered it. This ensures we only save each record in the chain once.